+++ src/main/java/ru/example/service/PassportService.java
@@ -5,8 +5,8 @@ import ru.example.model.Client;
 
 public interface PassportService {
 
-    PassportDTO findPassportByClientId(int clientId);
+    PassportDTO findPassportByClientId(long clientId);
     void savePassport(PassportDTO passportDTO, Client client);
     void updatePassport(PassportDTO passportDTO);
-    void deletePassportByClientId(int clientId);
+    void deletePassportByClientId(long clientId);
 }
